modified:   src/components/Header/Header.tsx     modified:   src/components/Input/Input.tsx 	renamed:    src/public/icons/logo.svg -> src/img/icons/logo.svg 	renamed:    src/public/icons/send.svg -> src/img/icons/send.svg 	renamed:    src/public/icons/voice.svg -> src/img/icons/voice.svg
